more incremental progress
