+++ rtx_package/src/engine/latex_delimiters.rs
@@ -6,58 +6,6 @@ use crate::prelude::*;
 // rather than the individual tokens.
 // This meaning is handled in MathParser (for now)
 
-/// A shorthand data structure for delimiter metadata
-pub struct DelimiterMeta {
-  char: char,
-  left_role: &'static str,
-  right_role: &'static str,
-  name: Option<&'static str>,
-
-}
-/// This duplicates in slightly different way what DefMath has put together.
-pub static DELIMITER_MAP : Lazy<HashMap<&'static str, DelimiterMeta>> = Lazy::new(|| raw_map!(
-  "(" => DelimiterMeta{char: '(', left_role: "OPEN", right_role: "CLOSE", name:None},
-  ")" => DelimiterMeta{char: ')', left_role: "OPEN", right_role: "CLOSE", name:None},
-  "[" => DelimiterMeta{char: '[', left_role: "OPEN", right_role: "CLOSE", name:None},
-  "]" => DelimiterMeta{ char: ']', left_role: "OPEN", right_role: "CLOSE", name:None},
-  "\\{" => DelimiterMeta{ char: '{', left_role: "OPEN", right_role: "CLOSE", name:None},
-  "\\}" => DelimiterMeta{ char: '}', left_role: "OPEN", right_role: "CLOSE", name:None},
-  "\\lfloor"=> DelimiterMeta{ char: '\u{230A}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("lfloor") },
-  "\\rfloor"=> DelimiterMeta{ char: '\u{230B}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("rfloor") },
-  "\\lceil" => DelimiterMeta{ char: '\u{2308}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("lceil") },
-  "\\rceil" => DelimiterMeta{ char: '\u{2309}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("rceil") },
-  "\\langle"=> DelimiterMeta{ char: '\u{27E8}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("langle") },
-  "\\rangle"=> DelimiterMeta{ char: '\u{27E9}',
-                left_role: "OPEN",  right_role: "CLOSE", name: Some("rangle") },
-  "<"      => DelimiterMeta{ char: '\u{27E8}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("langle") },
-  ">"      => DelimiterMeta{ char: '\u{27E9}',
-                left_role: "OPEN", right_role: "CLOSE", name: Some("rangle") },
-  "/"      => DelimiterMeta{ char: '/', left_role: "MULOP",   right_role: "MULOP", name: None },
-  "\\backslash" => DelimiterMeta{ char: '\u{005C}',
-                left_role: "MULOP",   right_role: "MULOP", name: Some("backslash") },
-  "|"      => DelimiterMeta{ char: '|',
-                left_role: "VERTBAR", right_role: "VERTBAR", name: None },
-  "\\|"     => DelimiterMeta{ char: '\u{2225}',
-                left_role: "VERTBAR", right_role: "VERTBAR", name: None },
-  "\\uparrow"   => DelimiterMeta{ char: '\u{2191}',
-                    left_role: "OPEN", right_role: "CLOSE", name: Some("uparrow") },
-  "\\Uparrow"   => DelimiterMeta{ char: '\u{21D1}',
-                    left_role: "OPEN", right_role: "CLOSE", name: Some("Uparrow") },
-  "\\downarrow" => DelimiterMeta{ char: '\u{2193}',
-                    left_role: "OPEN", right_role: "CLOSE", name: Some("downarrow") },
-  "\\Downarrow" =>  DelimiterMeta{ char: '\u{21D3}',
-                    left_role: "OPEN", right_role: "CLOSE", name: Some("Downarrow") },
-  "\\updownarrow" => DelimiterMeta{ char: '\u{2195}',
-                    left_role: "OPEN", right_role: "CLOSE", name: Some("updownarrow") },
-  "\\Updownarrow" => DelimiterMeta{ char: '\u{21D5}',
-                    left_role: "OPEN", right_role: "CLOSE", name: Some("Updownarrow") }
-));
 
 
 LoadDefinitions!({
@@ -103,81 +51,6 @@ LoadDefinitions!({
   //   my ($delim) = @_;
   //   return $DELIMITER_MAP{$delim}; }
 
-  // This is a little messier than you'd think.
-  // These effectively create a group between the \left,\right.
-  // And this also gives us a single list of things to parse separately.
-  // Since \left,\right are TeX, primitives and must be paired up,
-  // we use a bit of macro trickery to simulate.
-  // [The \@hidden@bgroup/egroup keep from putting a {} into the UnTeX]
-  // HOWEVER, an additional complication is that it is a common mistake to omit the balancing
-  // \right! Using an \egroup (or hidden) makes it hard to recover, so use a special egroup
-  DefMacro!("\\left XToken", r"\@left #1\@hidden@bgroup");
-  // Like \@hidden@egroup, but softer about missing \left
-  DefConstructor!("\\right@hidden@egroup", "",
-    after_digest => {
-      if is_value_bound("MODE", Some(0)) // Last stack frame was a mode switch!?!?!
-        || lookup_bool("groupNonBoxing") { // or group was opened with \begingroup
-        Error!("unexpected", "\\right", "Unbalanced \\right, no balancing \\left."); }
-      else {
-        egroup()?;
-      }
-    },
-    reversion => None);
-
-  DefMacro!("\\right XToken", r"\right@hidden@egroup\@right #1");
-
-  DefConstructor!("\\@left Token",
-    "?#char(<ltx:XMTok role='#role' name='#name' stretchy='#stretchy'>#char</ltx:XMTok>)\
-      (?#hint(<ltx:XMHint/>)(#1))",
-    after_digest => sub[whatsit] {
-      let delim = whatsit.get_arg(1).map(ToString::to_string).unwrap_or_default();
-      if delim == "." {
-        whatsit.set_property("hint", true); }
-      else if let Some(entry) = DELIMITER_MAP.get(delim.as_str()) {
-        whatsit.set_property("role", entry.left_role);
-        whatsit.set_property("char", entry.char);
-        whatsit.set_property("name", entry.name);
-        whatsit.set_property("stretchy", true);
-        // TODO: Should we have more Rc<> wrappers over Font?
-        whatsit.set_font(Rc::new(
-          whatsit.get_arg(1).unwrap().get_font()?.unwrap().into_owned()
-        ));
-      }
-      else if whatsit.get_arg(1).unwrap().get_property_string("role") == "OPEN" {
-        whatsit.get_arg_mut(1).unwrap().set_property("stretchy", true);
-      } else {
-        Warn!("unexpected", delim,
-          "Missing delimiter; '.' inserted");
-      }
-      Ok(Vec::new())
-    },
-    alias => "\\left");
-  DefConstructor!("\\@right Token",
-    "?#char(<ltx:XMTok role='#role' name='#name' stretchy='#stretchy'>#char</ltx:XMTok>)\
-      (?#hint(<ltx:XMHint/>)(#1))",
-    after_digest => sub[whatsit] {
-      let delim = whatsit.get_arg(1).map(ToString::to_string).unwrap_or_default();
-      if delim == "." {
-        whatsit.set_property("hint", true); }
-      else if let Some(entry) = DELIMITER_MAP.get(delim.as_str()) {
-        whatsit.set_property("role", entry.right_role);
-        whatsit.set_property("char", entry.char);
-        whatsit.set_property("name", entry.name);
-        whatsit.set_property("stretchy", true);
-        // TODO: Should we have more Rc<> wrappers over Font?
-        whatsit.set_font(Rc::new(
-          whatsit.get_arg(1).unwrap().get_font()?.unwrap().into_owned()
-        ));
-      }
-      else if whatsit.get_arg(1).unwrap().get_property_string("role") == "CLOSE" {
-        whatsit.get_arg_mut(1).unwrap().set_property("stretchy", true);
-      } else {
-        Warn!("unexpected", delim,
-          "Missing delimiter; '.' inserted");
-      }
-      Ok(Vec::new())
-    },
-    alias => "\\right");
 
   // These originally had Token as parameter, rather than {}..... Why?
   // Note that in TeX, \big{((} will only enlarge the 1st paren!!!
