+++ rtx_package/src/engine/tex_box.rs
@@ -149,8 +149,8 @@ LoadDefinitions!({
   AssignValue!("TEXT_MODE_BINDINGS"  => Stored::VecDequeStored(VecDeque::new()));
   AssignValue!("HTEXT_MODE_BINDINGS" => Stored::VecDequeStored(VecDeque::new()));
   AssignValue!("VTEXT_MODE_BINDINGS" => Stored::VecDequeStored(VecDeque::new()));
-  push_value("HTEXT_MODE_BINDINGS", Tokens!(T_MATH!(), T_CS!("\\@dollar@in@textmode")))?;
-  push_value("VTEXT_MODE_BINDINGS", Tokens!(T_MATH!(), T_CS!("\\@dollar@in@normalmode")))?;
+  push_value("HTEXT_MODE_BINDINGS", Tokens!(T_MATH!(), T_CS!("\\lx@dollar@in@textmode")))?;
+  push_value("VTEXT_MODE_BINDINGS", Tokens!(T_MATH!(), T_CS!("\\lx@dollar@in@normalmode")))?;
 
 
   // TODO: collapseSVGGroup
